Create README.md
# Octavo-Mandamiento
(wip project)
This is a WordPress project. It started out as a basic html,css and js site and I'm transferring it into a full fledged WordPress theme. The goal of the site is to be a News outlet in Nueva Esparta, Venezuela. 

I'm adding it to GitHub to log my progress and to be able to show my work to other colleges. I'll try to add issues and if anyone finds anything wrong, they can let me know as well!